pyproject.toml
[tool.poetry]
name = "flask-langchain-starter"
version = "0.1.0"
description = ""
authors = ["Tomas Rokos <tomas@rokos.cz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
langchain = "^0.2.1"
langchain-openai = "^0.1.7"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
